+++ sdk-java/src/test/java/ly/count/sdk/java/internal/TestUtils.java
@@ -1,0 +1,7 @@
+package ly.count.sdk.java.internal;
+
+public class TestUtils {
+    public String[] getCurrentRequestQueue() {
+        return new String[] {};
+    }
+}